% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneoverlap.R
\name{geneoverlap}
\alias{geneoverlap}
\title{Test if an overlap in gene selection methods occured at random}
\usage{
geneoverlap(N, d1, d2, d3)
}
\arguments{
\item{N}{a three-dimensional vector consisting of the number of genes used for selection in methods M1, M2, M3}

\item{d1}{a three-dimensional vector consisting of the number of genes selected by methods M1, M2, M3}

\item{d2}{a three-dimensional vector consisting of the number of genes selected by both methods (M1,M2), (M2,M3), (M3,M1)}

\item{d3}{the number of genes selected by all three methods (M1,M2,M3)}
}
\value{
p-value of the Z-score
}
\description{
Test if an overlap in gene selection methods occured at random
}
\details{
Suppose we have N genes and three methods M1, M2, M3 for
selecting a subset of genes, selecting each gene with some fixed probability p1, p2, p3 respectively.
This package tests if the intersection of these selection processes occured at random.
}
\examples{
N = c(34602,33627,22144)
d1 = c(458, 1938, 289) # A, B, C
d2 = c(105, 28, 172) # AB, BC, CA
d3 = c(47)

sign_overlap(N,d1,d2,d3)
}
